all permissionas and exceptions for subadmins and users are working

+++ app/Http/Controllers/NewsController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\News;
+use App\Models\User;
+use App\Models\Subadmin;
 use Illuminate\Http\Request;
 use Validator;
 use Carbon\Carbon;
@@ -19,25 +21,34 @@ class NewsController extends Controller
     public function __construct(){
         \Config::set('auth.defaults.guard', 'subadmin-api');
     }
-    public function index_subadmin()
-    {
-        if((\Auth::id()) && (\Auth::user()->role_type == 'subadmin')){
-            $new = News::where('subadmin_id',  auth()->user()->id)->get();
-            return response()->json(['news'=>$new],200);
-        }
-        else{
-            $news = News::all();    
-            return response()->json(['news'=>$news], 200);
-        }    
-            
-    }
 
+    // Display only for useers 
     public function index_user()
     {
         $news = News::all();    
         return response()->json(['news'=>$news], 200);      
     }
 
+    public function show_user($id)
+    {
+        $news = News::find($id);
+        if(!$news){
+            return response()->json(['message' => 'Not Found!'],404);
+        }
+
+        return response()->json(['news'=>$news], 200);
+
+    
+    }
+
+    // Display only for specific subadmins
+    
+    public function index_subadmin()
+    {   
+        $new = News::where('subadmin_id',  auth()->user()->id)->get();
+        return response()->json(['news'=>$new],200);
+    }
+   
     public function store(NewsStoreRequest $request)
     {
        try{
@@ -49,7 +60,7 @@ class NewsController extends Controller
                 'title_of_news' => $request->title_of_news,
                 'description_of_news' => $request->description_of_news,
                 'photo_of_news' => $photo,
-                'date_of_news' => $request->date_of_news
+                'date_of_news' => Carbon::create($request->date_of_news)->toDateString()
              
             ]);
 
@@ -62,14 +73,18 @@ class NewsController extends Controller
         }
     }
 
-    public function show($id)
+    public function show_subadmin($id)
     {
-        $news = News::find($id);
-        if($news){
-            return response()->json(['news'=>$news], 200);
+       //$news = News::where('subadmin_id', \Auth::id())->where('id', $id)->first();
+        $news = News::find($id);    
+        if(!$news){
+            return response()->json(['message'=> 'Not Found!'],404);
+        }
+        elseif($news->subadmin_id == auth()->user()->id){
+             return response()->json(['message'=> $news],200);
         }
         else{
-            return response()->json(['message' => 'Not Found!'],404);
+            return response()->json(['message'=>"U can see only ur's news! . Look better your id's in show reoute!"],200);
         }
     }
 
@@ -81,29 +96,33 @@ class NewsController extends Controller
             if(!$news){
                 return response()->json(['message' => 'Not Found!'],404);
             }
+            elseif($news->subadmin_id == auth()->user()->id){
+                
+                $news->title_of_news = $request->title_of_news;
+                $news->description_of_news = $request->description_of_news;
+                $news->date_of_news = Carbon::create($request->date_of_news)->toDateString();
 
-            $news->title_of_news = $request->title_of_news;
-            $news->description_of_news = $request->description_of_news;
-            $news->date_of_news = $request->date_of_news;
+                if($request->photo_of_news){
+                    $storage = Storage::disk('public'); //public storage
 
-            if($request->photo_of_news){
-                $storage = Storage::disk('public'); //public storage
+                    //old image delete
+                    if($storage->exists($news->photo_of_news))
+                        $storage->delete($news->photo_of_news);
 
-                //old image delete
-                if($storage->exists($news->photo_of_news))
-                    $storage->delete($news->photo_of_news);
+                    //Image name
+                    $photo = Str::random(32).".".$request->photo_of_news->getClientOriginalExtension();
+                    $news->photo_of_news = $photo;
 
-                //Image name
-                $photo = Str::random(32).".".$request->photo_of_news->getClientOriginalExtension();
-                $news->photo_of_news = $photo;
+                    //Image save in public folder
+                    $storage->put($photo, file_get_contents($request->photo_of_news));
+                }
 
-                //Image save in public folder
-                $storage->put($photo, file_get_contents($request->photo_of_news));
+                $news->save();
+                return response()->json(['message' => 'News successfully updated!'], 200);
             }
-
-            $news->save();
-            return response()->json(['message' => 'News successfully updated!'], 200);
-
+            else{
+                return response()->json(['message' => "U can't update! ,It's not ur's news. U can edit only urs news!"],200);
+            }
         
         } catch(\Exception $e){
             return response()->json(['message' => 'Something went wrong!'],500);
@@ -117,13 +136,18 @@ class NewsController extends Controller
         if(!$news){
             return response()->json(['message' => 'Not Found!'], 404);
         }
+        elseif($news->subadmin_id == auth()->user()->id){
         
-        $storage = Storage::disk('public');
-        if($storage->exists($news->photo_of_news))
-            $storage->delete($news->photo_of_news);
+            $storage = Storage::disk('public');
+            if($storage->exists($news->photo_of_news))
+                $storage->delete($news->photo_of_news);
 
-        $news->delete();
+            $news->delete();
 
-        return response()->json(['message' => 'News successfully deleted!'], 200);
+            return response()->json(['message' => 'News successfully deleted!'], 200);
+        }
+        else{
+            return response()->json(['message' =>"U can't delete news, Delete only urs new"], 200);
+        }
     }
 }
